+++ oci/mounts.go
@@ -68,3 +68,6 @@ func defaultMounts() []specs.Mount {
 		},
 	}
 }
+
+// appendOSMounts is only used on FreeBSD, and a no-op on other platforms.
+func appendOSMounts(_ *Spec, _ string) {}
